requirements.txt
flask==3.1.0
aiohttp==3.11.12
gunicorn==23.0.0
mysql-connector-python==9.2.0
flask-mysql==1.6.0
flask-sqlalchemy==3.1.1
python-dotenv==1.0.1